requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
Flask==3.0.2
Flask-SocketIO==5.3.6
Pillow>=10.0.0
psutil
tiktoken
playwright
watchdog
chromadb
